+++ server/impl/src/main/java/com/walmartlabs/concord/server/websocket/WebSocketCreator.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.server.websocket;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -70,8 +70,9 @@ public class WebSocketCreator implements org.eclipse.jetty.websocket.servlet.Web
         }
 
         UUID channelId = UUID.randomUUID();
-        String channelInfo = req.getHeader(InternalConstants.Headers.AGENT);
-        return new WebSocketListener(channelManager, channelId, channelInfo);
+        String agentId = req.getHeader(InternalConstants.Headers.AGENT_ID);
+        String userAgent = req.getHeader(InternalConstants.Headers.AGENT_UA);
+        return new WebSocketListener(channelManager, channelId, agentId, userAgent);
     }
 
     private static boolean invalidApiKey(String s) {